+++ mne_bids/report.py
@@ -13,9 +13,9 @@ from mne.utils import warn, logger, verbose
 
 from mne_bids.config import DOI, ALLOWED_DATATYPES
 from mne_bids.tsv_handler import _from_tsv
-from mne_bids.path import (get_datatypes, get_entity_vals, BIDSPath,
-                           _parse_ext, _find_matching_sidecar,
-                           get_entities_from_fname)
+from mne_bids.path import (get_bids_path_from_fname, get_datatypes,
+                           get_entity_vals, BIDSPath,
+                           _parse_ext, _find_matching_sidecar)
 
 # functions to be used inside Template strings
 FUNCTION_TEMPLATE = """{{py:
@@ -325,8 +325,9 @@ def _summarize_sidecar_json(root, scans_fpaths):
             n_scans += 1
 
             # convert to BIDS Path
-            params = get_entities_from_fname(bids_path)
-            bids_path = BIDSPath(root=root, **params)
+            if not isinstance(bids_path, BIDSPath):
+                bids_path = get_bids_path_from_fname(bids_path)
+            bids_path.root = root
 
             # XXX: improve to allow emptyroom
             if bids_path.subject == 'emptyroom':
@@ -410,8 +411,9 @@ def _summarize_channels_tsv(root, scans_fpaths):
                 continue
 
             # convert to BIDS Path
-            params = get_entities_from_fname(bids_path)
-            bids_path = BIDSPath(root=root, **params)
+            if not isinstance(bids_path, BIDSPath):
+                bids_path = get_bids_path_from_fname(bids_path)
+            bids_path.root = root
 
             # XXX: improve to allow emptyroom
             if bids_path.subject == 'emptyroom':